change an example from the manual (#14)
Currently the tests are run for the GAP versions from 4.12 to dev.
In GAP 4.15, new `Irr` methods (may) become available,
which affects the ordering of irreducibles in character tables.

Now we avoid `Display` in manual examples.

(First I tried to get an example in whoch the same `Irr` method
is used for the GAP versions in question.
The best such method for supersolvable groups is the Baum-Clausen
based one, but this does not store power maps in older GAP versions.
For the nonsolvable group A5,
apparently the orderings obtained in GAP 4.12 and GAP 4.15 are also
different.)